# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amrutam Telemedicine API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/doctors": {
      "get": {
        "tags": [
          "doctors"
        ],
        "summary": "List Doctors",
        "operationId": "list_doctors_v1_doctors_get",
        "parameters": [
          {
            "name": "specialization",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Specialization"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorOut"
                  },
                  "title": "Response List Doctors V1 Doctors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/doctors/slots": {
      "post": {
        "tags": [
          "doctors"
        ],
        "summary": "Create Slot",
        "operationId": "create_slot_v1_doctors_slots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/doctors/{doctor_id}/slots": {
      "get": {
        "tags": [
          "doctors"
        ],
        "summary": "List Slots",
        "operationId": "list_slots_v1_doctors__doctor_id__slots_get",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doctor Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SlotOut"
                  },
                  "title": "Response List Slots V1 Doctors  Doctor Id  Slots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/consultations/{cid}/prescriptions": {
      "post": {
        "tags": [
          "prescriptions"
        ],
        "summary": "Create Rx",
        "operationId": "create_rx_v1_consultations__cid__prescriptions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/prescriptions/{rx_id}": {
      "get": {
        "tags": [
          "prescriptions"
        ],
        "summary": "View Rx",
        "operationId": "view_rx_v1_prescriptions__rx_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rx_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rx Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/prescriptions/{rx_id}/pdf": {
      "get": {
        "tags": [
          "prescriptions"
        ],
        "summary": "Rx Pdf",
        "operationId": "rx_pdf_v1_prescriptions__rx_id__pdf_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rx_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rx Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfUrl"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Create Payment",
        "operationId": "create_payment_v1_payments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/payments/webhook": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Webhook",
        "operationId": "webhook_v1_payments_webhook_post",
        "parameters": [
          {
            "name": "X-Webhook-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Webhook-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/search/doctors": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search Doctors",
        "operationId": "search_doctors_v1_search_doctors_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "specialization",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Specialization"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorOut"
                  },
                  "title": "Response Search Doctors V1 Search Doctors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DoctorOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "specialization": {
            "type": "string",
            "title": "Specialization"
          }
        },
        "type": "object",
        "required": [
          "id",
          "specialization"
        ],
        "title": "DoctorOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentCreate": {
        "properties": {
          "consultation_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Consultation Id"
          },
          "amount": {
            "type": "integer",
            "maximum": 1000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "maxLength": 8,
            "title": "Currency",
            "default": "INR"
          },
          "force_fail": {
            "type": "boolean",
            "title": "Force Fail",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "consultation_id",
          "amount"
        ],
        "title": "PaymentCreate"
      },
      "PaymentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "consultation_id": {
            "type": "string",
            "title": "Consultation Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "gateway_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gateway Ref"
          }
        },
        "type": "object",
        "required": [
          "id",
          "consultation_id",
          "status"
        ],
        "title": "PaymentOut"
      },
      "PdfUrl": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_in_s": {
            "type": "integer",
            "title": "Expires In S",
            "default": 300
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "PdfUrl"
      },
      "PrescriptionCreate": {
        "properties": {
          "diagnosis": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Diagnosis"
          },
          "medicines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Medicines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "diagnosis",
          "medicines"
        ],
        "title": "PrescriptionCreate"
      },
      "PrescriptionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "consultation_id": {
            "type": "string",
            "title": "Consultation Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "consultation_id"
        ],
        "title": "PrescriptionOut"
      },
      "PrescriptionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "consultation_id": {
            "type": "string",
            "title": "Consultation Id"
          },
          "diagnosis": {
            "type": "string",
            "title": "Diagnosis"
          },
          "medicines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Medicines"
          }
        },
        "type": "object",
        "required": [
          "id",
          "consultation_id",
          "diagnosis",
          "medicines"
        ],
        "title": "PrescriptionView"
      },
...
      "SlotCreate": {
        "properties": {
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "starts_at",
          "ends_at"
        ],
        "title": "SlotCreate"
      },
      "SlotOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "doctor_id": {
            "type": "string",
            "title": "Doctor Id"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "doctor_id",
          "starts_at",
          "ends_at",
          "status",
          "version"
        ],
        "title": "SlotOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookPayload": {
        "properties": {
          "consultation_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Consultation Id"
          },
          "status": {
            "type": "string",
            "pattern": "^(paid|failed)$",
            "title": "Status"
          },
          "gateway_ref": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Gateway Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "consultation_id",
          "status",
          "gateway_ref"
        ],
        "title": "WebhookPayload"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}